pyproject.toml
[project]
name = "customer-churn-prediction"
version = "0.1.0"
description = "A machine learning project to predict customer churn"
authors = [
    {name = "Maxwell Selassie"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.3.2,<3.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "python-box (>=7.3.2,<8.0.0)",
    "ensure (>=1.0.4,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"

